refactor(LC038): split eager-loading analysis helpers
- separate chain counting, ancestor detection, and threshold resolution
- keep analyzer behavior unchanged while reducing local complexity
